fix: reduce cycles
