+++ app/client/src/sagas/DebuggerSagas.ts
@@ -58,8 +58,8 @@ import {
 import { identifyEntityFromPath } from "../navigation/FocusEntity";
 import { getIDEViewMode } from "../selectors/ideSelectors";
 import type { EditorViewMode } from "IDE/Interfaces/EditorTypes";
-import { getDebuggerPaneConfig } from "../components/editorComponents/Debugger/hooks/useDebuggerConfig";
 import { DEBUGGER_TAB_KEYS } from "../components/editorComponents/Debugger/constants";
+import { getDebuggerPaneConfig } from "../components/editorComponents/Debugger/utils/getDebuggerPaneConfig";
 
 let blockedSource: string | null = null;
